project-playwright/pyproject.toml
[project]
name = "project-playwright"
version = "0.1.0"
description = "Playwright 漸進式學習專案 — 從零開始掌握瀏覽器自動化與資料擷取"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"

dependencies = [
    "playwright>=1.49.0",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-playwright>=0.6.2",
    "pytest-asyncio>=0.24.0",
]
scraping = [
    "pandas>=2.2.0",
    "openpyxl>=3.1.0",
]
all = [
    "project-playwright[dev,scraping]",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.pytest.ini_options]
testpaths = ["ch07-testing"]
markers = [
    "slow: 需要較長時間的測試",
    "network: 需要網路連線的測試",
]
